css/css-transforms/animation/rotate-interpolation-002.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>rotate interpolation</title>
<link rel="help" href="https://drafts.csswg.org/css-transforms-2/#propdef-rotate">

<meta name="assert" content="rotate supports animation.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>
</head>

<style>
.parent {
  rotate: z 90deg;
}

.target {
  width: 40px;
  height: 20px;
  background-color: grey;
  rotate: 10deg;
}

.expected {
  background-color: green;
}
</style>

<template id="target-template">
<div class="parent">
    <div class="target">Text</div>
</div>
</template>

<body>
<script>
test_interpolation({
  property: 'rotate',
  from: 'inherit',
  to: '270deg',
}, [
  {at: -1, expect: 'z -90deg'},
  {at: 0, expect: 'z 90deg'},
  {at: 0.25, expect: 'z 135deg'},
  {at: 0.75, expect: 'z 225deg'},
  {at: 1, expect: 'z 270deg'},
  {at: 2, expect: 'z 450deg'},
]);
</script>
</body>
</html>
